move auth middleware to revisar utils
