+++ revisar-server-utils/index.ts
@@ -1,4 +1,4 @@
 export { default as createLogger } from './logger';
 export { default as initDB } from './db';
-export { schemaValidationMiddleware } from './middlewares';
+export { schemaValidationMiddleware, authMiddleware} from './middlewares';
 export { createServer } from './server';
